further deconstruction of homepage / more components
