+++ src/Components/ProgramInfo.css
@@ -1,0 +1,7 @@
+.miles {
+    padding-top: 15px;
+}
+
+.acts {
+    padding-top: 15px;
+}
